Add files via upload
Combined TA and FireRecord datasets to make 'fire/no_fire' column where the dates are equal. To elaborate, the TA Pandas Dataframe has dates and thermal anomaly values, and the FireRecord dataset has dates of fires (initially ranges but Kaylee Pham flattened it to singular dates at a time). The 4th column of the TA dataset (classifying fire or no fire, 'Class') was initially set to no fire, and now is set to fire when the dates aligned in both datasets.