Created one side function from main + relative path dir feature
Broke one of the preexisting cells into a function and added function 'create_abs_path_to_dataset()' that sets up an abs path to the dataset when dataset dir name is passed in. This should simplify setup steps when running on different local machines as it eliminates need for changing abs dir every time it's downloaded. Note: the dataset should be in the same directory as the .py file running (e.g. both in dataset_by_county folder)